[wip] add android-build-aar step and test
