+++ tools/expotools/src/versioning/android/android-packages-to-keep.txt
@@ -1,0 +1,3 @@
+expo.modules.gl.cpp
+org.unimodules.interfaces.taskManager
+org.unimodules.core.interfaces.SingletonModule
