+++ tools/expotools/src/versioning/android/android-packages-to-rename.txt
@@ -1,0 +1,15 @@
+com.facebook.catalyst
+com.facebook.csslayout
+com.facebook.jni
+com.facebook.perftest
+com.facebook.proguard
+com.facebook.quicklog
+com.facebook.react
+com.facebook.systrace
+com.facebook.yoga
+com.facebook.debug
+org.unimodules.core
+org.unimodules.interfaces
+expo.modules
+org.unimodules.adapters
+org.unimodules.errors
