--- tools/expotools/src/versioning/android/constants.ts
@@ -1,42 +1,0 @@
-export const JniLibNames = [
-  'libfb',
-  'libfbjni',
-  'libfolly_json',
-  'libglog_init',
-  'glog',
-  'reactnativejni',
-  'reactnativejnifb',
-  'csslayout',
-  'yoga',
-  'fbgloginit',
-  'yogajni',
-  'jschelpers',
-  'packagerconnectionjnifb',
-  'privatedata',
-  'yogafastmath',
-  'fabricjscjni',
-  'jscexecutor',
-  'libjscexecutor',
-  'jsinspector',
-  'libjsinspector',
-  'fabricjni',
-  'turbomodulejsijni',
-];
-
-export const JavaPackagesToRename = [
-  'com.facebook.catalyst',
-  'com.facebook.csslayout',
-  'com.facebook.jni',
-  'com.facebook.perftest',
-  'com.facebook.proguard',
-  'com.facebook.quicklog',
-  'com.facebook.react',
-  'com.facebook.systrace',
-  'com.facebook.yoga',
-  'com.facebook.debug',
-  'org.unimodules.core',
-  'org.unimodules.interfaces',
-  'expo.modules',
-  'org.unimodules.adapters',
-  'org.unimodules.errors'
-];
