+++ tools/expotools/src/versioning/android/template-stripe-activity.xml
@@ -1,0 +1,14 @@
+		<!-- BEGIN_SDK_$ABI_VERSION_NUMBER -->
+		<activity
+			android:exported="true"
+			android:launchMode="singleTask"
+			android:name="abi$ABI_VERSION_NUMBER.expo.modules.payments.stripe.RedirectUriReceiver"
+			android:theme="@android:style/Theme.Translucent.NoTitleBar.Fullscreen">
+			<intent-filter>
+				<action android:name="android.intent.action.VIEW" />
+					<category android:name="android.intent.category.DEFAULT" />
+					<category android:name="android.intent.category.BROWSABLE" />
+					<data android:scheme="abi$ABI_VERSION_NUMBER.expo.modules.payments.stripe" />
+			</intent-filter>
+		</activity>
+		<!-- END_SDK_$ABI_VERSION_NUMBER -->
